move 'Entry' to list.rs
